Separate routers into v1 and v2
