+++ src/routers/v2/RouteRouter.js
@@ -1,11 +1,11 @@
 // @flow
 import type Request from 'express';
-import AnalyticsUtils from '../utils/AnalyticsUtils';
-import ApplicationRouter from '../appdev/ApplicationRouter';
-import LogUtils from '../utils/LogUtils';
-import SectionRouteUtils from '../utils/SectionRouteUtils';
+import AnalyticsUtils from '../../utils/AnalyticsUtils';
+import ApplicationRouter from '../../appdev/ApplicationRouter';
+import LogUtils from '../../utils/LogUtils';
+import RouteV2Utils from '../../utils/RouteV2Utils';
 
-class SectionRouteRouter extends ApplicationRouter<Array<Object>> {
+class RouteRouter extends ApplicationRouter<Array<Object>> {
   constructor() {
     super(['POST']);
   }
@@ -16,18 +16,18 @@ class SectionRouteRouter extends ApplicationRouter<Array<Object>> {
 
   async content(req: Request): Promise<Object> {
     const {
-      isArriveBy,
       destinationName,
       end,
+      isArriveBy,
       originName,
       start,
       time,
       uid,
     } = req.body;
 
-    const isOriginBusStop = await SectionRouteUtils.isBusStop(originName);
+    const isOriginBusStop = await RouteV2Utils.isBusStop(originName);
     const originBusStopName = isOriginBusStop ? originName : null;
-    const sectionedRoutes = await SectionRouteUtils.getSectionedRoutes(
+    const sectionedRoutes = await RouteV2Utils.getSectionedRoutes(
       destinationName,
       end,
       start,
@@ -35,7 +35,7 @@ class SectionRouteRouter extends ApplicationRouter<Array<Object>> {
       isArriveBy,
       originBusStopName,
     );
-    const routes = SectionRouteUtils.flatten(Object.values(sectionedRoutes));
+    const routes = RouteV2Utils.flatten(Object.values(sectionedRoutes));
     if (routes.length > 0) {
       const request = {
         isArriveBy,
@@ -55,4 +55,4 @@ class SectionRouteRouter extends ApplicationRouter<Array<Object>> {
   }
 }
 
-export default new SectionRouteRouter().router;
+export default new RouteRouter().router;
